refactor: move build command to be a sub section from program section and remove installation processes
