+++ crates/scilla/src/commands/main_command.rs
@@ -3,12 +3,11 @@ use {
         commands::{Command, CommandFlow, navigation::NavigationSection},
         context::ScillaContext,
         prompt::{
-            prompt_account_section, prompt_build_section, prompt_cluster_section,
-            prompt_config_section, prompt_program_section, prompt_stake_section,
-            prompt_transaction_section, prompt_vote_section,
+            prompt_account_section, prompt_cluster_section, prompt_config_section,
+            prompt_program_section, prompt_stake_section, prompt_transaction_section,
+            prompt_vote_section,
         },
     },
-    anyhow::Ok,
     std::fmt,
 };
 
@@ -19,7 +18,6 @@ pub enum MainCommand {
     Program,
     Vote,
     Transaction,
-    Build,
     ScillaConfig,
     Exit,
 }
@@ -34,7 +32,6 @@ impl fmt::Display for MainCommand {
             MainCommand::Vote => "Vote",
             MainCommand::Transaction => "Transaction",
             MainCommand::ScillaConfig => "Scilla Config",
-            MainCommand::Build => "Build",
             MainCommand::Exit => "Exit",
         };
         f.write_str(label)
@@ -53,7 +50,6 @@ impl Command for MainCommand {
             MainCommand::Vote => prompt_vote_section()?.process_command(ctx).await?,
             MainCommand::Transaction => prompt_transaction_section()?.process_command(ctx).await?,
             MainCommand::Program => prompt_program_section()?.process_command(ctx).await?,
-            MainCommand::Build => prompt_build_section()?.process_command(ctx).await?,
             MainCommand::ScillaConfig => prompt_config_section()?.process_command(ctx).await?,
             MainCommand::Exit => {
                 return Ok(CommandFlow::Exit);
