+++ crates/scilla/src/commands/mod.rs
@@ -5,7 +5,6 @@ use {
 };
 
 pub mod account;
-pub mod build;
 pub mod cluster;
 pub mod config;
 pub mod main_command;
