+++ crates/scilla/src/commands/program/mod.rs
@@ -7,11 +7,10 @@ use {
         context::ScillaContext,
         prompt::{prompt_go_back, prompt_program_section_shared},
     },
-    anyhow::Ok,
     core::fmt,
 };
 
-mod build;
+pub mod build;
 mod close;
 mod deploy;
 mod extend;
@@ -63,14 +62,14 @@ impl fmt::Display for ProgramShared {
 }
 
 impl ProgramShared {
-    fn process_command(&self, _ctx: &ScillaContext) -> CommandFlow {
+    async fn process_command(&self, ctx: &mut ScillaContext) -> anyhow::Result<CommandFlow> {
         match self {
             ProgramShared::Deploy => todo!(),
             ProgramShared::Upgrade => todo!(),
-            ProgramShared::Build => todo!(),
+            ProgramShared::Build => build::process_build(ctx).await,
             ProgramShared::Close => todo!(),
             ProgramShared::Extend => todo!(),
-            ProgramShared::GoBack => CommandFlow::NavigateTo(prompt_go_back()),
+            ProgramShared::GoBack => Ok(CommandFlow::NavigateTo(prompt_go_back())),
         }
     }
 }
@@ -106,13 +105,13 @@ impl Command for ProgramCommand {
                 ctx.get_nav_context_mut()
                     .checked_push(NavigationSection::ProgramLegacy);
                 let command = prompt_program_section_shared()?;
-                command.process_command(ctx)
+                command.process_command(ctx).await?
             }
             ProgramCommand::ProgramV4 => {
                 ctx.get_nav_context_mut()
                     .checked_push(NavigationSection::ProgramV4);
                 let command = prompt_program_section_shared()?;
-                command.process_command(ctx)
+                command.process_command(ctx).await?
             }
             ProgramCommand::GoBack => {
                 return Ok(CommandFlow::NavigateTo(NavigationTarget::PreviousSection));
