+++ crates/scilla/src/prompt.rs
@@ -3,12 +3,11 @@ use {
         commands::{
             Command,
             account::AccountCommand,
-            build::{BuildCommand, BuildMode},
             cluster::ClusterCommand,
             config::ConfigCommand,
             main_command::MainCommand,
             navigation::NavigationTarget,
-            program::{ProgramCommand, ProgramShared},
+            program::{ProgramCommand, ProgramShared, build::BuildMode},
             stake::StakeCommand,
             transaction::TransactionCommand,
             vote::VoteCommand,
@@ -32,7 +31,6 @@ pub fn prompt_main_section() -> anyhow::Result<impl Command> {
             MainCommand::Program,
             MainCommand::Vote,
             MainCommand::Transaction,
-            MainCommand::Build,
             MainCommand::ScillaConfig,
             MainCommand::Exit,
         ],
@@ -110,17 +108,6 @@ pub fn prompt_program_section() -> anyhow::Result<ProgramCommand> {
             ProgramCommand::ProgramV4,
             ProgramCommand::GoBack,
         ],
-    )
-    .with_page_size(10)
-    .prompt()?;
-
-    Ok(choice)
-}
-
-pub fn prompt_build_section() -> anyhow::Result<BuildCommand> {
-    let choice = Select::new(
-        "Build Command:",
-        vec![BuildCommand::Build, BuildCommand::GoBack],
     )
     .with_page_size(10)
     .prompt()?;
@@ -281,13 +268,6 @@ pub fn prompt_keypair_path(msg: &str, ctx: &ScillaContext) -> PathBuf {
 
 pub fn prompt_confirmation(msg: &str) -> bool {
     Confirm::new(msg).prompt().unwrap_or(false)
-}
-
-pub fn prompt_confirm_install(msg: &str) -> bool {
-    Confirm::new(msg)
-        .with_default(true)
-        .prompt()
-        .unwrap_or(false)
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
